introduce cache in StatisticsSummary get method

+++ src/main/scala/com/despegar/metrik/model/HistogramBucket.scala
@@ -32,7 +32,8 @@ class HistogramBucket(override val bucketNumber: Long, override val duration: Du
     val max = histogram.getMaxValue
     val count = histogram.getTotalCount
     val avg = histogram.getMean
-    StatisticSummary(timestamp, p50, p80, p90, p95, p99, p999, min, max, count, avg)
+
+    StatisticSummary(timestamp, p50, p80, p90, p95, p99, p999, min, max, count, avg.toLong)
   }
 }
